src/function-apps/frontend/requirements.txt
azure-functions
cryptography
azure-identity
azure-core
msal
requests
setuptools-rust